+++ core/libs/facesengine/recognition/opencv-dnn/kd_tree.cpp
@@ -81,7 +81,7 @@ KDNode* KDTree::add(const cv::Mat& position, const int identity)
     return newNode;
 }
 
-QMap<double, QVector<int> > KDTree::getClosestNeighbors(const cv::Mat& position, double sqRange, int maxNbNeighbors) const
+QMap<double, QVector<int> > KDTree::getClosestNeighbors(const cv::Mat& position, float sqRange, int maxNbNeighbors) const
 {
     // TODO : verify if we need a mutex for reading
